Separting NFTHeader from AccountDetailsPage, adding a new TokenSection for NFTs and adding some cleanup
